% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glue_ann.R
\name{glue_ann}
\alias{glue_ann}
\title{Train a GLUE-ANN model ensemble}
\usage{
glue_ann(inputData, outputData, orthogonal = "Cor",
  variableNumber = c(1:ncol(inputData)), dimRed = T, rescale.output = T,
  ES = T, randomESPart = 0.5, nonEqualMeans = T,
  maxMeanDiff = mean(outputData) * 0.2, nCycles = 100, HLrange = c(1:1),
  HNrange = c(1:10), maxHNPerVar = Inf, hidden.layer = "tansig",
  output.layer = "sigmoid", nSets = 100, nStedinger = 50,
  weighting = "Stedinger", cv = F)
}
\arguments{
\item{inputData}{input data frame}

\item{outputData}{output data frame}

\item{orthogonal}{character; orthogonalize the input data; options are: 'Cov', 'Cor', and 'No'; default: 'Cor'}

\item{variableNumber}{number of input variables; provide vector for stochastic treatment; defaults to number of columns of inputData}

\item{dimRed}{logical; TRUE selects always most variance explaining PCs; FALSE chooses PCs randomly; default: TRUE}

\item{rescale.output}{logical; rescale output to [0,1] range; default: TRUE}

\item{ES}{logical; enable early stopping; default: TRUE}

\item{randomESPart}{part of the data used for early stopping; default: 0.5}

\item{nonEqualMeans}{force similar mean for the early stopping samples as for the training samples; default: TRUE}

\item{maxMeanDiff}{maximum difference between early stopping and training observation means; defaults to 20 percent of the target data mean}

\item{nCycles}{maximum number of training cycles, provide a vector for stochastic treatment; default: 100}

\item{HLrange}{amount of hidden layers, provide vector for stochastic treatment; default: c(1:1)}

\item{HNrange}{amount of hidden nodes, provide vector for stochastic treatment; default: c(1:10)}

\item{maxHNPerVar}{maximum amount of hidden nodes per input variable; default: Inf}

\item{hidden.layer}{hidden layer activation function; options are 'tansig', 'sigmoid', 'purelin'; default: 'tansig'}

\item{output.layer}{output layer activation function; options are 'tansig', 'sigmoid', 'purelin'; default: 'sigmoid'}

\item{nSets}{number of models in the ensemble; default: 100}

\item{nStedinger}{number of random variates for weights based on Stedinger et al. (2008); default: 50}

\item{weighting}{GLUE weighting procedure; options are 'MSE', 'MSEes' (for weights based on the early stopping subset performance), 'NSEff' and 'Stedinger'; default: 'Stedinger'}

\item{cv}{logical; do leave-one-out cross-validation; default: FALSE}
}
\value{
Object of class glue_ann
}
\description{
\code{glue_ann} trains a GLUE-ANN model ensemble. A dataframe has to be provided with the independent variable(s), and a vector with the dependent variable. A list of parameters has to be set to define the a priori distributions of the stochastic parameters, and fixed values for the ones considered to be constant.
}

