\name{nncalc}
\alias{nncalc}
\title{ANN training function}
\usage{
nncalc(Ptrain, Pval, targetTrain, targetVal, architecture, n, Stao = 1,
  es = T, hidden.layer = "sigmoid", output.layer = "purelin")
}
\arguments{
  \item{Ptrain}{}

  \item{Pval}{}

  \item{targetTrain}{}

  \item{targetVal}{}

  \item{architecture}{}

  \item{n}{}

  \item{Stao}{}

  \item{es}{}

  \item{hidden.layer}{}

  \item{output.layer}{}
}
\value{
results of training an ANN
}
\description{
\code{nncalc} is a wrapper for the AMORE package functions
}

