% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.glue_ann}
\alias{predict.glue_ann}
\title{Make predictions with a GLUE-ANN model ensemble}
\usage{
\method{predict}{glue_ann}(glue_ann, newdata,
  weighting = glue_ann$parameters$weighting)
}
\arguments{
\item{glue_ann}{a glue_ann object}

\item{newdata}{input data frame for making predictions}

\item{weighting}{GLUE weighting procedure; options are 'MSE', 'MSEes' (for weights based on the early stopping subset performance), 'NSEff' and 'Stedinger'; default: 'Stedinger'}
}
\value{
predictions from glue_ann model ensemble
}
\description{
\code{predict.glue_ann} makes predictions for newdata with a GLUE-ANN model ensemble
}

