% Generated by roxygen2 (4.0.2): do not edit by hand
\name{train_ann}
\alias{train_ann}
\title{ANN training function}
\usage{
train_ann(Ptrain, Pval, targetTrain, targetVal, architecture, n, Stao = 1,
  es = T, hidden.layer = "sigmoid", output.layer = "purelin")
}
\arguments{
\item{Ptrain}{}

\item{Pval}{}

\item{targetTrain}{}

\item{targetVal}{}

\item{architecture}{}

\item{n}{}

\item{Stao}{}

\item{es}{}

\item{hidden.layer}{}

\item{output.layer}{}
}
\value{
results of training an ANN
}
\description{
\code{train_ann} is a wrapper for the AMORE package functions
}

